Refactor drum processing (#609)
* make pitch offset type consistent

* add pitch offset to note tracker and note events

* refactor drum node

* remove unused sample id field and update related functions

* formatting

+++ shared/src/model/track.rs
@@ -29,7 +29,7 @@ pub struct PlacedNote {
     #[proto_optional]
     pub note: Note,
     pub offset: OrderedFloat<Beats>,
-    pub pitch_offset: i32,
+    pub pitch_offset: f32,
 }
 
 impl PartialOrd for PlacedNote {
